+++ tests/offline/test_offline_plates.R
@@ -5,7 +5,7 @@ model <- chronosphere::fetch("paleomap", "model", datadir=file.path(wd, "data/ch
 
 ################################################################################
 # today
-rec0 <- reconstruct("plates", 0, model=model)
+rec0 <- reconstruct("static_polygons", 0, model=model)
 
 
 # correct type
@@ -17,7 +17,7 @@ expect_equal(sf::st_crs(rec0)[[1]], "WGS 84")
 
 ################################################################################
 # past
-rec140 <- reconstruct("plates", 140, model=model)
+rec140 <- reconstruct("static_polygons", 140, model=model)
 
 
 expect_inherits(rec140, "sf")
@@ -25,7 +25,7 @@ expect_equal(sf::st_crs(rec140)[[1]], "WGS 84")
 
 ################################################################################
 # Multiple
-rec <- reconstruct("plates", age=c(0, 140), model=model)
+rec <- reconstruct("static_polygons", age=c(0, 140), model=model)
 
 
 expect_inherits(rec, "list")