+++ tests/online/test_online_plates.R
@@ -6,7 +6,7 @@ options(timeout = 5*60)
 ################################################################################
 # today
 expect_silent(
-	rec0 <- reconstruct("plates", 0, model="PALEOMAP")
+	rec0 <- reconstruct("static_polygons", 0, model="MERDITH2021")
 )
 
 # correct type
@@ -19,7 +19,7 @@ expect_equal(sf::st_crs(rec0)[[1]], "4326")
 ################################################################################
 # past
 expect_silent(
-	rec140 <- reconstruct("plates", 140, model="PALEOMAP")
+	rec140 <- reconstruct("static_polygons", 140, model="MERDITH2021")
 )
 
 expect_inherits(rec140, "sf")
@@ -28,7 +28,7 @@ expect_equal(sf::st_crs(rec140)[[1]], "4326")
 ################################################################################
 # Multiple
 expect_silent(
-	rec <- reconstruct("plates", age=c(0, 140), model="PALEOMAP")
+	rec <- reconstruct("static_polygons", age=c(0, 140), model="MERDITH2021")
 )
 
 expect_inherits(rec, "list")
@@ -40,7 +40,7 @@ expect_equal(rec[["140"]], rec140)
 # Test inaccurate ages! - only rounded ones are allowed
 # Reconstructed to 140.4 
 expect_message(
-	rec140_4 <- reconstruct("plates", age=140.4, model="PALEOMAP")
+	rec140_4 <- reconstruct("static_polygons", age=140.4, model="MERDITH2021")
 )	
 
 expect_equal(rec140_4, rec140)
@@ -49,35 +49,35 @@ expect_equal(rec140_4, rec140)
 # Multiple models
 # Seton model
 expect_silent(
-	rec <- reconstruct("plates", 0, model="SETON2012")
+	rec <- reconstruct("static_polygons", 0, model="SETON2012")
 )
 expect_inherits(rec, "sf")
 expect_equal(sf::st_crs(rec)[[1]], "4326")
 
 # Müller 2016
 expect_silent(
-	rec <- reconstruct("plates", 0, model="MULLER2016")
+	rec <- reconstruct("static_polygons", 0, model="MULLER2016")
 )
 expect_inherits(rec, "sf")
 expect_equal(sf::st_crs(rec)[[1]], "4326")
 
 # the GOLONKA model
 expect_silent(
-	rec <- reconstruct("plates", 0, model="GOLONKA")
+	rec <- reconstruct("static_polygons", 0, model="GOLONKA")
 )
 expect_inherits(rec, "sf")
 expect_equal(sf::st_crs(rec)[[1]], "4326")
 
 # the scotese model
-expect_silent(
-	rec <- reconstruct("plates", 0, model="PALEOMAP")
-)
-expect_inherits(rec, "sf")
-expect_equal(sf::st_crs(rec)[[1]], "4326")
+## expect_silent(
+## 	rec <- reconstruct("static_polygons", 0, model="PALEOMAP")
+## )
+## expect_inherits(rec, "sf")
+## expect_equal(sf::st_crs(rec)[[1]], "4326")
 
 # Matthew 2016
 expect_silent(
-	rec <- reconstruct("plates", 0, model="MATTHEWS2016")
+	rec <- reconstruct("static_polygons", 0, model="MATTHEWS2016")
 )
 expect_inherits(rec, "sf")
 expect_equal(sf::st_crs(rec)[[1]], "4326")
@@ -85,6 +85,6 @@ expect_equal(sf::st_crs(rec)[[1]], "4326")
 # invalid input
 suppressWarnings(
 	expect_error(
-		rec <- reconstruct("plates", 0, model="GARBAGE")
+		rec <- reconstruct("static_polygons", 0, model="GARBAGE")
 	)
 )